@@ -1,7 +1,7 @@
 package com.example.app;
 
-import javax.ws.rs.ApplicationPath;
-import javax.ws.rs.core.Application;
+import jakarta.ws.rs.ApplicationPath;
+import jakarta.ws.rs.core.Application;
 
 @ApplicationPath(value = "/")
 public class WebApplication extends Application {
