@@ -1,6 +1,6 @@
 package com.example.app.com;
 
-import javax.enterprise.context.RequestScoped;
+import jakarta.enterprise.context.RequestScoped;
 
 @RequestScoped
 public class RequestContext {
