@@ -3,8 +3,8 @@ package com.example.app.com.servlet.filter;
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
 
-import javax.servlet.ReadListener;
-import javax.servlet.ServletInputStream;
+import jakarta.servlet.ReadListener;
+import jakarta.servlet.ServletInputStream;
 
 public class BufferedServletInputStream extends ServletInputStream {
 
