@@ -1,7 +1,7 @@
 package com.example.app.rest.service;
 
-import javax.enterprise.context.ApplicationScoped;
-import javax.inject.Inject;
+import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.inject.Inject;
 
 @ApplicationScoped
 public class DecoratorSuper implements DecoratorIF {
