@@ -1,6 +1,6 @@
 package com.example.app.rest.service;
 
-import javax.enterprise.context.ApplicationScoped;
+import jakarta.enterprise.context.ApplicationScoped;
 
 @ApplicationScoped
 public class Samplehogehge {
