@@ -1,22 +1,22 @@
 package com.example.app.soa;
 
-import javax.enterprise.context.ApplicationScoped;
+import jakarta.enterprise.context.ApplicationScoped;
 
 @ApplicationScoped
 public class SoaSample {
 
 	public void sample() {
-		SimpleAddBean service = new SimpleAddBean();
-		SimpleAddBeanPortType port = service.getSimpleAddBeanPort();
-		String reply = port.getSample();
-		
+//		SimpleAddBean service = new SimpleAddBean();
+//		SimpleAddBeanPortType port = service.getSimpleAddBeanPort();
+//		String reply = port.getSample();
+//		
 //		JaxWsProxyFactoryBean factory = new JaxWsProxyFactoryBean(); 
 //		factory.setServiceClass(SimpleAddBeanPortType.class);
 //		factory.setHandlers(Arrays.asList(new SampleSoapHandler()));
 //		factory.setAddress("http://localhost:9090/SimpleAddBeanPort"); 
 //		SimpleAddBeanPortType client = (SimpleAddBeanPortType) factory.create();
 //		String reply = client.getSample(); 
-		System.out.println("Server said: " + reply); 
+//		System.out.println("Server said: " + reply);
 	}
-	
+
 }
