@@ -1,7 +1,7 @@
 package com.example.app.com.jaxrs.exceptionmapper.handler;
 
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
+import jakarta.ws.rs.core.MediaType;
+import jakarta.ws.rs.core.Response;
 
 public class AppBusinessExceptionHandlerTest {
 	
