@@ -2,9 +2,9 @@ package com.example.app.soa;
 
 import java.util.List;
 
-import javax.xml.ws.handler.Handler;
-import javax.xml.ws.handler.HandlerResolver;
-import javax.xml.ws.handler.PortInfo;
+import jakarta.xml.ws.handler.Handler;
+import jakarta.xml.ws.handler.HandlerResolver;
+import jakarta.xml.ws.handler.PortInfo;
 
 public class SoaHandlerResolver implements HandlerResolver {
 
