@@ -2,20 +2,18 @@ package com.example.app;
 
 import org.junit.Rule;
 
-import com.example.app.soa.SimpleAddBean;
-import com.example.app.soa.SimpleAddBeanPortType;
 import com.github.tomakehurst.wiremock.junit.WireMockClassRule;
 
 public class WireMockTest {
 
 	@Rule
 	public WireMockClassRule wiremockClassRule = new WireMockClassRule(9090);
-	
+
 	public void test() {
-		SimpleAddBean service = new SimpleAddBean();
-		SimpleAddBeanPortType port = service.getSimpleAddBeanPort();
-		String reply = port.getSample();
-		System.out.println(reply);
+//		SimpleAddBean service = new SimpleAddBean();
+//		SimpleAddBeanPortType port = service.getSimpleAddBeanPort();
+//		String reply = port.getSample();
+//		System.out.println(reply);
 //		JaxWsProxyFactoryBean factory = new JaxWsProxyFactoryBean(); 
 //		factory.setServiceClass(SimpleAddBeanPortType.class);
 //		factory.setHandlers(Arrays.asList(new SampleSoapHandler()));
@@ -24,5 +22,5 @@ public class WireMockTest {
 //		String reply = client.getSample(); 
 
 	}
-	
+
 }
